Move methods for test
