+++ install/build.py
@@ -12,6 +12,52 @@ minimum_cuda_version = 6050
 minimum_cudnn_version = 2000
 
 
+def get_compiler_setting():
+    nvcc_path = utils.search_on_path(('nvcc', 'nvcc.exe'))
+    cuda_path_default = None
+    if nvcc_path is None:
+        utils.print_warning('nvcc not in path.',
+                            'Please set path to nvcc.')
+    else:
+        cuda_path_default = os.path.normpath(
+            os.path.join(os.path.dirname(nvcc_path), '..'))
+
+    cuda_path = os.environ.get('CUDA_PATH', '')  # Nvidia default on Windows
+    if len(cuda_path) > 0 and cuda_path != cuda_path_default:
+        utils.print_warning(
+            'nvcc path != CUDA_PATH',
+            'nvcc path: %s' % cuda_path_default,
+            'CUDA_PATH: %s' % cuda_path)
+
+    if not os.path.exists(cuda_path):
+        cuda_path = cuda_path_default
+
+    if not cuda_path and os.path.exists('/usr/local/cuda'):
+        cuda_path = '/usr/local/cuda'
+
+    include_dirs = []
+    library_dirs = []
+    define_macros = []
+
+    if cuda_path:
+        include_dirs.append(os.path.join(cuda_path, 'include'))
+        if sys.platform == 'win32':
+            library_dirs.append(os.path.join(cuda_path, 'bin'))
+            library_dirs.append(os.path.join(cuda_path, 'lib', 'x64'))
+        else:
+            library_dirs.append(os.path.join(cuda_path, 'lib64'))
+            library_dirs.append(os.path.join(cuda_path, 'lib'))
+    if sys.platform == 'darwin':
+        library_dirs.append('/usr/local/cuda/lib')
+
+    return {
+        'include_dirs': include_dirs,
+        'library_dirs': library_dirs,
+        'define_macros': define_macros,
+        'language': 'c++',
+    }
+
+
 def check_cuda_version(compiler, settings):
     try:
         out = build_and_run(compiler, '''
